requirements.txt
# ========== Core Trading ==========
alpaca-py>=0.21.0
python-telegram-bot>=20.0
requests>=2.31.0

# ========== Data Processing ==========
pandas>=2.0.0
numpy>=1.24.0

# ========== Portfolio Optimization ==========
pyportfolioopt>=1.5.5  # Black-Litterman, Mean-Variance Optimization
cvxpy>=1.4.0           # 凸优化求解器 (pypfopt 依赖)

# ========== LLM / AI Agent ==========
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph-checkpoint-postgres>=2.0.0  # PostgreSQL checkpointer + store
# langchain-deepseek>=0.1.0  # 可选：DeepSeek 模型支持

# ========== Web Framework ==========
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9  # FastAPI 表单解析

# ========== Database ==========
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0  # SQLite 异步驱动
psycopg[binary,pool]>=3.1.0  # PostgreSQL 驱动 (langgraph-checkpoint-postgres 依赖)

# ========== Authentication ==========
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# ========== Scheduling & Async ==========
apscheduler>=3.10.0
tenacity>=8.2.0
aiolimiter>=1.1.0
aiohttp>=3.9.0
websockets>=12.0

# ========== Time & Calendar ==========
pytz>=2024.1
exchange-calendars>=4.5.0  # 交易日历（节假日、交易时间）

# ========== Configuration ==========
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ========== Logging ==========
structlog>=24.1.0

# ========== Testing ==========
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
